+++ src/main.rs
@@ -1,30 +1,25 @@
 //! Entry point.
 
 use crate::db::Db;
-use crate::message::*;
+use crate::prelude::*;
 use crate::settings::Settings;
 use crossbeam_channel::{Receiver, Sender};
-use failure::{format_err, Error};
-use log::{debug, info, warn, Level};
+use log::Level;
 use std::path::PathBuf;
 use std::sync::{Arc, Mutex};
 use structopt::StructOpt;
 
-// TODO: perhaps remove `pub`.
-pub mod consts;
-pub mod core;
-pub mod db;
-pub mod format;
-pub mod message;
-pub mod persistence;
-pub mod services;
-pub mod settings;
-pub mod supervisor;
-pub mod templates;
-pub mod value;
-pub mod web;
-
-type Result<T> = std::result::Result<T, Error>;
+mod consts;
+mod core;
+mod db;
+mod format;
+mod persistence;
+mod prelude;
+mod services;
+mod settings;
+mod supervisor;
+mod templates;
+mod web;
 
 #[derive(StructOpt, Debug)]
 #[structopt(name = "my-iot", author, about)]
@@ -70,7 +65,7 @@ fn main() -> Result<()> {
     // - the dispatcher sends out each message from `dispatcher_rx` to the services input channels
     info!("Starting services…");
     let (dispatcher_tx, dispatcher_rx) = crossbeam_channel::unbounded();
-    dispatcher_tx.send(Composer::new("my-iot::start").type_(Type::ReadNonLogged).into())?;
+    dispatcher_tx.send(Composer::new("my-iot::start").type_(MessageType::ReadNonLogged).into())?;
     let mut all_txs = vec![persistence::spawn(db.clone(), &dispatcher_tx)?];
     all_txs.extend(spawn_services(&settings, &db, &dispatcher_tx)?);
     spawn_dispatcher(dispatcher_rx, dispatcher_tx, all_txs)?;