+++ src/persistence.rs
@@ -1,9 +1,8 @@
 //! Readings persistence.
 
 use crate::db::*;
-use crate::message::*;
+use crate::prelude::*;
 use crate::supervisor;
-use crate::Result;
 use crossbeam_channel::Sender;
 use log::{debug, info};
 use std::sync::{Arc, Mutex};
@@ -28,10 +27,11 @@ pub fn spawn(db: Arc<Mutex<Db>>, tx: &Sender<Message>) -> Result<Sender<Message>
 fn process_message(message: Message, db: &Arc<Mutex<Db>>, tx: &Sender<Message>) -> Result<()> {
     info!("{}: {:?} {:?}", &message.sensor, &message.type_, &message.value,);
     debug!("{:?}", &message);
-    if message.type_ == Type::ReadLogged {
+    // TODO: handle `ReadSnapshot`.
+    if message.type_ == MessageType::ReadLogged {
         let db = db.lock().unwrap();
         let previous_reading = db.select_last_reading(&message.sensor)?;
-        db.insert_reading(&message)?;
+        db.upsert_reading(&message)?;
         send_messages(&previous_reading, &message, &tx)?;
     }
     Ok(())
@@ -43,14 +43,14 @@ fn send_messages(previous_reading: &Option<Message>, message: &Message, tx: &Sen
         if message.timestamp > existing.timestamp {
             tx.send(
                 Composer::new(format!("{}::update", &message.sensor))
-                    .type_(Type::ReadNonLogged)
+                    .type_(MessageType::ReadNonLogged)
                     .value(message.value.clone())
                     .into(),
             )?;
             if message.value != existing.value {
                 tx.send(
                     Composer::new(format!("{}::change", &message.sensor))
-                        .type_(Type::ReadNonLogged)
+                        .type_(MessageType::ReadNonLogged)
                         .value(message.value.clone())
                         .into(),
                 )?;