+++ src/services/clock.rs
@@ -1,7 +1,5 @@
-use crate::message::*;
+use crate::prelude::*;
 use crate::supervisor;
-use crate::value::Value;
-use crate::Result;
 use crossbeam_channel::Sender;
 use serde::Deserialize;
 use std::thread;